requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==4.2.2
    # via
    #   audio-recorder-streamlit
    #   streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
audio-recorder-streamlit==0.0.10
    # via noetia (pyproject.toml)
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   openai-chat
    #   streamlit
colorama==0.4.6
    # via
    #   click
    #   openai-chat
    #   tqdm
contourpy==1.3.3
    # via matplotlib
cryptography==46.0.5
    # via google-auth
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
easyocr==1.7.2
    # via noetia (pyproject.toml)
entrypoints==0.4
    # via altair
filelock==3.24.3
    # via torch
fonttools==4.61.1
    # via matplotlib
fsspec==2026.2.0
    # via torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-auth==2.48.0
    # via google-auth-oauthlib
google-auth-oauthlib==1.2.4
    # via noetia (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
imageio==2.37.2
    # via scikit-image
imageio-ffmpeg==0.6.0
    # via noetia (pyproject.toml)
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
jiter==0.13.0
    # via openai
joblib==1.5.3
    # via scikit-learn
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.8
    # via seaborn
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via
    #   scikit-image
    #   torch
ninja==1.13.0
    # via easyocr
numpy==2.4.2
    # via
    #   altair
    #   contourpy
    #   easyocr
    #   imageio
    #   matplotlib
    #   opencv-python-headless
    #   pandas
    #   pydeck
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   seaborn
    #   shapely
    #   streamlit
    #   tifffile
    #   torchvision
    #   xgboost
oauthlib==3.3.1
    # via requests-oauthlib
openai==2.21.0
    # via
    #   noetia (pyproject.toml)
    #   openai-chat
openai-chat==0.0.3
    # via noetia (pyproject.toml)
opencv-python-headless==4.13.0.92
    # via easyocr
packaging==26.0
    # via
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   streamlit
pandas==2.3.3
    # via
    #   altair
    #   seaborn
    #   streamlit
pillow==12.1.1
    # via
    #   easyocr
    #   imageio
    #   matplotlib
    #   scikit-image
    #   streamlit
    #   torchvision
prompt-toolkit==3.0.52
    # via openai-chat
protobuf==6.33.5
    # via streamlit
pyarrow==23.0.1
    # via streamlit
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pyclipper==1.4.0
    # via easyocr
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyparsing==3.3.2
    # via matplotlib
python-bidi==0.6.7
    # via easyocr
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.2.1
    # via noetia (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via easyocr
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   openai-chat
    #   requests-oauthlib
    #   streamlit
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
scikit-image==0.26.0
    # via easyocr
scikit-learn==1.8.0
    # via noetia (pyproject.toml)
scipy==1.17.0
    # via
    #   easyocr
    #   scikit-image
    #   scikit-learn
    #   xgboost
seaborn==0.13.2
    # via noetia (pyproject.toml)
setuptools==82.0.0
    # via torch
shapely==2.1.2
    # via easyocr
six==1.17.0
    # via
    #   python-dateutil
    #   whisper
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via openai
streamlit==1.54.0
    # via
    #   noetia (pyproject.toml)
    #   audio-recorder-streamlit
    #   streamlit-calendar
streamlit-calendar==1.4.0
    # via noetia (pyproject.toml)
sympy==1.14.0
    # via torch
tenacity==9.1.4
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2026.2.20
    # via scikit-image
toml==0.10.2
    # via streamlit
toolz==1.1.0
    # via altair
torch==2.10.0
    # via
    #   easyocr
    #   torchvision
torchvision==0.25.0
    # via easyocr
tornado==6.5.4
    # via streamlit
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
wcwidth==0.6.0
    # via prompt-toolkit
whisper==1.1.10
    # via noetia (pyproject.toml)
xgboost==3.2.0
    # via noetia (pyproject.toml)
